rename and adjusted the drops for cage
