--- src/main/java/jamdoggie/betterbattletowers/block/BlockLogicNonSolid.java
@@ -1,27 +1,0 @@
-package jamdoggie.betterbattletowers.block;
-
-import net.minecraft.core.block.Block;
-import net.minecraft.core.block.BlockLogic;
-import net.minecraft.core.block.material.Material;
-import net.minecraft.core.world.WorldSource;
-
-public class BlockLogicNonSolid extends BlockLogic {
-	public BlockLogicNonSolid(Block<?> block, Material material) {
-		super(block, material);
-	}
-
-	@Override
-	public boolean isSolidRender() {
-		return false;
-	}
-
-	@Override
-	public boolean renderAsNormalBlockOnCondition(WorldSource world, int x, int y, int z) {
-		return false;
-	}
-
-	@Override
-	public boolean blocksLight() {
-		return true;
-	}
-}
